Decouple WebView and WebView client form the business logic.
Added WebViewCoordinator which provide the connection between the client and the view. Custom business logic must extend this coordinator.
Disable overriding the KolibriWebViewClient. Developers must set listener instead.
WebViewCoordinator is example for custom logic depending on parsing the HTML meta and request headers. From now this is not responsibility of the view and the web client.